Add Hide Not Completed toggle
Adds an Hide Not Completed toggle in the Sorting menu. When turned on this hides all anime that haven't aired yet or are currently airing. This might mis some since the method for retrieving the air date is broken for anime that aren't airing, so it checks the total episodes count, and air date for airing anime.